base: move vformat.h
