+++ src/v/base/BUILD
@@ -12,12 +12,14 @@ redpanda_cc_library(
         "include/base/type_traits.h",
         "include/base/units.h",
         "include/base/vassert.h",
+        "include/base/vformat.h",
         "include/base/vlog.h",
     ],
     strip_include_prefix = "include",
     visibility = ["//visibility:public"],
     deps = [
         "@boost//:outcome",
+        "@fmt",
         "@seastar",
     ],
 )
