+++ src/v/base/include/base/vformat.h
